+++ Configs/InventoryManagement.cs
@@ -110,7 +110,7 @@ public sealed class PreviousSlot {
     [DefaultValue(true)] public bool consumption = true;
     [DefaultValue(true)] public bool mediumCore = true;
     [DefaultValue(false)] public bool overridePrevious = false;
-    [DefaultValue(true)] public bool moveItems = true;
+    [DefaultValue(MovePolicy.NotFavorited)] public MovePolicy movePolicy = MovePolicy.NotFavorited;
     public Toggle<Materials> materials = new(true);
     public Toggle<PreviousDisplay> displayPrevious = new(true);
 
@@ -120,10 +120,12 @@ public sealed class PreviousSlot {
     public static PreviousSlot Value => SmartPickup.Value.previousSlot.Value;
 }
 
+public enum MovePolicy { Never, NotFavorited, Always }
+
 public sealed class QuickStackPickup {
     [DefaultValue(true)] public bool voidBag = true;
     [DefaultValue(true)] public bool chests = true;
-    
+
     public static QuickStackPickup Value => SmartPickup.Value.quickStack.Value;
 }