+++ Utility.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Runtime.CompilerServices;
+using BetterInventory.InventoryManagement;
 using BetterInventory.InventoryManagement.SmartPickup;
 using MonoMod.Cil;
 using SpikysLib;
@@ -38,24 +39,24 @@ public static class Utility {
 
     public static int FailedILs { get; private set; }
 
-    public static Item? LastStack(this Player player, Item item, AllowedItems allowedItems = AllowedItems.None) {
-        bool Check(Item i) => item.type == i.type && (allowedItems.HasFlag(AllowedItems.Self) || i != item);
+    public static Item? LastStack(this Player player, Item item, StackPickerSettings settings) {
+        bool Check(Item i) => item.type == i.type && (settings.CanPickArg || i != item);
 
         for (int i = InventorySlots.Items.End - 1; i >= InventorySlots.Items.Start; i--) if (Check(player.inventory[i])) return player.inventory[i];
         for (int i = InventorySlots.Ammo.End - 1; i >= InventorySlots.Coins.Start; i--) if (Check(player.inventory[i])) return player.inventory[i];
-        if (allowedItems.HasFlag(AllowedItems.Mouse) && Check(player.inventory[InventorySlots.Mouse])) return player.inventory[InventorySlots.Mouse];
+        if (settings.CanPickMouse && Check(player.inventory[InventorySlots.Mouse])) return player.inventory[InventorySlots.Mouse];
         return null;
     }
 
-    public static Item? SmallestStack(this Player player, Item item, AllowedItems allowedItems = AllowedItems.None) {
+    public static Item? SmallestStack(this Player player, Item item, StackPickerSettings settings) {
         Item? min = null;
         void Check(Item i) {
-            if (item.type == i.type && (min is null || i.stack < min.stack) && (allowedItems.HasFlag(AllowedItems.Self) || i != item)) min = i;
+            if (item.type == i.type && (min is null || i.stack < min.stack) && (settings.CanPickArg || i != item)) min = i;
         }
 
         for (int i = InventorySlots.Items.End - 1; i >= InventorySlots.Items.Start; i--) Check(player.inventory[i]);
         for (int i = InventorySlots.Ammo.End - 1; i >= InventorySlots.Coins.Start; i--) Check(player.inventory[i]);
-        if (allowedItems.HasFlag(AllowedItems.Mouse)) Check(player.inventory[InventorySlots.Mouse]);
+        if (settings.CanPickMouse) Check(player.inventory[InventorySlots.Mouse]);
         return min;
     }